requirements.txt
python-dotenv
langchain==0.1.0
langchain_openai
pydantic==1.10.8
docarray
openai
pygments
